refactor : MVC 패턴으로 패키지를 변경하다
