+++ src/main/java/Service/Calculator.java
@@ -1,4 +1,4 @@
-package domain;
+package Service;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
